Rename the profiles db file to a name that better indicates that it's a db
